Enable FIPS endpoints for ecr-login
FIPS endpoints take the form ecr-fips.<region>.<domain suffix>.

NOTE: Using a FIPs endpoint with ECR requires passing in "ecr-fips" as the
service name to resolve to the correct client endpoint, but we do not
expose "ecr-fips" as a recognized service name distinct from "ecr".
Ignoring the error allows the sdk to still return the correct endpoint.