debian: remove client_test.go
The mocks in client_test.go depend on the ecriface.ECRAPI interface, as
the client.go code uses that for accessing the ECR API client.  This
interface comes from the AWS SDK for Go, which is provided by the
golang-github-aws-aws-sdk-go-dev package in Debian.  The interface in
that package can change as new functionality is added to the ECR API and
released in the SDK.  This makes the mocks no longer match the
interface, and causes compilation to fail.

This patch removes the client_test.go file.  Upstream issue
https://github.com/awslabs/amazon-ecr-credential-helper/issues/145
tracks the work required to resolve the issue correctly.

See: https://github.com/awslabs/amazon-ecr-credential-helper/pull/136#issuecomment-455395713
See: https://github.com/awslabs/amazon-ecr-credential-helper/issues/145