api: remove dependency on ecriface package
The ecriface.ECRAPI interface presents an interface for the entire ECR
API.  This is useful if you use the entire surface of the API, but the
credential helper only uses a single API call: GetAuthorizationToken.
The use of the ecriface.ECRAPI interface for generating mock objects
meant that any time the interface changed, the mock objects needed to be
regenerated.  This has caused difficulty with upgrading the AWS SDK
(both when vendored, and when using a non-vendored SDK package like is
done in Debian).

This commit moves to a very small interface that contains the single
function that the credential helper calls, and regenerates the mock
based on that small interface.

Closes https://github.com/awslabs/amazon-ecr-credential-helper/issues/145