.codebuild: scripts for release pipeline
This commit sets up buildspecs for two different AWS CodeBuild builds,
intended to be run from AWS CodePipeline.  The main buildspec builds all
variants of the binary, and can also be run directly in AWS CodeBuild
with no changes.  The second buildspec produces correct source archives
that include commit information but not a complete git repository, as
well as metadata files like a version-specific section of the
CHANGELOG.md.