Use the XDG Basedir Spec for the cache
Currently, for caching data, a new directory is created in the current
user's home directory. This is bad practice (if all apps just dump files
into the user's home, it becomes somewhat unmaintainable).

The XDG Basedir Specification covers this pretty neatly; the cache
directory for all applications is configurable via the `XDG_CACHE_HOME`
variable, and falls back to `~/.cache` if unset.

This changeset makes the cache for all applications live in one single
location, rather that littering the user's home.