version: rename default version
The default version is used when this repository is built from source
using a default `go build` command.  The primary development branch has
been renamed from "master" to "main", but using the branch name as the
version isn't super clear.  The new "development" default version makes
it clear that the binary was built from source.