Don't require amazonaws.com in ECR url pattern
Some deployments of AWS are served behind a proxy or have an alternative domain name. The current code checks for `amazonaws.com` in the domain name when it gets matches via the regex (see https://github.com/awslabs/amazon-ecr-credential-helper/blob/main/ecr-login/api/client.go#L40). The domain name does not appear to be used in the matches argument and could be removed to enable using this against proxy, alternative, or iso region deployments of AWS.