Adds back Dockerfile for local builds
This commit reverts the removal of the repositories Dockerfile for local
builds. The container was helpful for users which need to build the
credential helper from source but do not have an environment to do so.

In addition to the revert, this chance adds dependabot configuration to
update the base image for the container and reworks the docker make
target to build artifacts with correct user permissions.

Signed-off-by: Austin Vazquez <macedonv@amazon.com>